requirements.txt
altgraph==0.17.4
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.2
firebase-admin==6.8.0
Flask==3.1.0
Flask-SocketIO==5.5.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.0.0
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.32.3
rsa==4.9.1
setuptools==80.1.0
simple-websocket==1.1.0
sseclient==0.0.27
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
finnhub-python
rich
pandas
eventlet
psycopg2
flask_login
pydantic[email]
fastapi
guvicorn
python-multipart